Disable HTTP2 for REST service calls
This is a test. Debugging some issue related to
matrix-corporal + Apache interaction over HTTP2,
which causes `Transfer-Encoding: chunked` to be used for no good reason.

Using chunked transfer encoding with HTTP2 is not supported, so it's
odd that it's happening.

Forcing HTTP 1.1 to see if that helps the Apache server receive the
payload and it that is indeed our issue.